+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the `C-mantic` extension will be documented in this file.
 ## [Unreleased]
 ### Added
 - Added `Update Function Signature` code-action. After changing a function signature, this code-action will be provided to update that function's declaration or definition.
+- Added a setting `Code Actions: Enable Update Function Signature`. When disabled, C-mantic will not spend any time tracking changes to function signatures.
 - Added completion suggestions to `Add Include`.
 
 ### Changed